Adds preview support for iOS 6 integrated auth to FBSession
Summary:
This diff does the following things -
  1. Introduces a define that will compile-in preview support for iOS 6 integration
  2. Updates the samples to include the Accounts package
  3. Minor tweaks to two of the samples

Some notes -
  * The stability of the iOS 6 integration is not 100%, so the code includes some
    odd defensive measures, including adding a permission when the list is nil, etc.
  * The integration of iOS 6 is still provisional and there are known changes, such
    as support for integrated caching, which are not present here

Test Plan: Ran each sample with integrated and non integrated auth, significant whitebox testing as well, ran a subset of the samples using 4.3 simulator to confirm fallback behavior

Reviewers: clang, mmarucheck, gregschechte

Reviewed By: clang

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D513350

Task ID: 1142879

+++ src/FBTestSession.m
@@ -348,7 +348,7 @@ tokenCachingStrategy:(FBSessionTokenCachingStrategy*)tokenCachingStrategy
     // We know each character is a digit. Convert it into a letter starting with 'a'.
     for (int i = 0; i < hashAsString.length; ++i) {
         NSString *ch = [NSString stringWithFormat:@"%C",
-                        [hashAsString characterAtIndex:i] + 'a' - '0'];
+                        (unsigned short)([hashAsString characterAtIndex:i] + 'a' - '0')];
         [result appendString:ch];
     }
